credit-score-system/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
